落英秘典

+++ Core/GObject/World.h
@@ -734,7 +734,7 @@ public:
     inline static void setSysDailogPlatform(UInt8 v) { m_sysDailogPlatform = v; }
     inline static UInt8 getSysDailogPlatform() { return m_sysDailogPlatform; }
     static Player* getSpreadKeeper();
-    static UInt16 getSpreadCount();
+    static UInt32 getSpreadCount();
 public:
     static UInt32 _moneyLogged;
     static MoneyIn _moneyIn[7][2];
@@ -854,7 +854,7 @@ public:
     static UInt8 _snakespringequipact;
     static UInt8 m_sysDailogPlatform;
     static Player* spreadKeeper;
-    static UInt16 spreadCount;
+    static UInt32 spreadCount;
 public:
     static RCSortType rechargeSort;
     static RCSortType consumeSort;
